# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zenith Alpha Capital API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /investment-plans/{id}:
    get:
      operationId: getInvestmentPlan
      tags: [investments]
      summary: Get investment plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentPlan"

  # ── Portfolios (active investments) ────────────────────────────────────────
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List user transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, profit, bonus, referral]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Request deposit or withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Admin ───────────────────────────────────────────────────────────────────
...
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user status/role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/transactions:
    get:
      operationId: listAllTransactions
      tags: [admin]
      summary: List all transactions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /admin/transactions/{id}/approve:
    post:
      operationId: approveTransaction
      tags: [admin]
      summary: Approve a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/transactions/{id}/reject:
    post:
      operationId: rejectTransaction
      tags: [admin]
      summary: Reject a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/investment-plans:
    post:
      operationId: createInvestmentPlan
      tags: [admin]
      summary: Create an investment plan (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentPlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentPlan"
  /admin/investment-plans/{id}:
    patch:
      operationId: updateInvestmentPlan
      tags: [admin]
      summary: Update an investment plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentPlanUpdate"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentPlan"
    delete:
      operationId: deleteInvestmentPlan
      tags: [admin]
      summary: Delete an investment plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Crypto Prices ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, email, role, status, balance, totalInvested, totalProfit, referralCode, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, pending]
        balance:
          type: number
        totalInvested:
          type: number
        totalProfit:
          type: number
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        createdAt:
          type: string
...
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, pending]
        balance:
          type: number
    InvestmentPlan:
      type: object
      required: [id, name, description, minAmount, maxAmount, returnRate, durationDays, riskLevel, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: ["number", "null"]
        returnRate:
          type: number
        durationDays:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high]
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    InvestmentPlanInput:
      type: object
      required: [name, description, minAmount, returnRate, durationDays, riskLevel]
      properties:
        name:
          type: string
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: ["number", "null"]
        returnRate:
          type: number
        durationDays:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high]
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    InvestmentPlanUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: ["number", "null"]
        returnRate:
          type: number
        durationDays:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high]
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
    Transaction:
      type: object
      required: [id, userId, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, profit, bonus, referral]
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        description:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        rejectReason:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        paymentMethod:
          type: string
        walletAddress:
          type: string
        description:
          type: string
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...
    BlogPost:
      type: object
      required: [id, title, slug, excerpt, content, authorName, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        authorName:
          type: string
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        createdAt:
          type: string
    BlogPostInput:
      type: object
      required: [title, slug, excerpt, content, authorName]
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImage:
          type: string
        authorName:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImage:
          type: string
        authorName:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
...