# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zenith Alpha Capital API
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from './userKycStatus';
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  role: UserRole;
  status: UserStatus;
  balance: number;
  totalInvested: number;
  totalProfit: number;
  referralCode: string;
  /** @nullable */
  referredBy?: string | null;
  kycStatus?: UserKycStatus;
  createdAt: string;
}